pyproject.toml
[tool.poetry]
name = "python_backend_template"
version = "0.1.0"
description = ""
authors = ["Admin <backend@napoleonit.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "1.10.13"}
# pydantic-settings = "^2.0.3"
sentry-sdk = "^1.9.10"
pyhumps = "^3.7.2"
dependency-injector = "^4.41.0"
structlog = "^22.1.0"
orjson = "^3.7.8"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
opentelemetry-sdk = "^1.13.0"
opentelemetry-instrumentation-fastapi = "^0.34b0"
opentelemetry-instrumentation-logging = "^0.34b0"
protobuf = "3.20.1"
opentelemetry-exporter-jaeger = "^1.13.0"
opentelemetry-propagator-jaeger = "^1.14.0"
opentelemetry-opentracing-shim = "^0.34b0"
hvac = "^2.0.0"
# httpx-backoff = "^1.0.0" #TODO: should add another backoff library
opentelemetry-instrumentation-httpx = "^0.34b0"


faust-streaming = "^0.8.11"
python-schema-registry-client = {extras = ["faust"], version = "^2.4.1"}


async_keycloak = "0.28.6"


sqlalchemy = {extras = ["asyncio"], version = "^1.4.41"}
greenlet = "^1.1.3"
sqlalchemy-stubs = "^0.4"
alembic = "^1.8.0"
asyncpg = "^0.26.0"
opentelemetry-instrumentation-sqlalchemy = "^0.34b0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
better-exceptions = "^0.3.3"
yappi = "^1.3.6"
trio = "^0.21.0"
pytest = "^7.1.3"
pytest-xdist = "^2.5.0"
pytest-sugar = "^0.9.5"
pytest-factoryboy = "^2.5.0"
pytest-instafail = "^0.4.2"
pytest-cov = "^3.0.0"
pytest-trio = "^0.7.0"
pytest-freezegun = "^0.4.2"
pytest-async-sqlalchemy = "~0.2.0"
pytest-lazy-fixture = "^0.6.3"
mypy = "^0.971"

[[tool.poetry.source]]
name = "itnap"
url = "https://nexus.itnap.ru/repository/pypi-proxy/simple"
default = true
secondary = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
python_files = "tests.py test_*.py *_tests.py"
addopts = "--numprocesses=0"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \S*/tests/\S*
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
